web_scraper/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
beautifulsoup4==4.11.1
bs4==0.0.1
requests==2.29.0
pandas==2.0.1
azure-storage-blob==12.16.0